requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
apispec==6.6.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bech32==1.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
cosmpy==0.9.2
dataclasses-json==0.6.6
distro==1.9.0
ecdsa==0.18.0
exceptiongroup==1.2.1
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.60
MarkupSafe==2.1.5
marshmallow==3.21.2
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging==23.2
protobuf==4.25.3
pycryptodome==3.20.0
pydantic==1.10.15
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.1
requests-file==2.1.0
rpds-py==0.18.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.7.0
tldextract==5.1.2
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
uagents==0.11.1
urllib3==2.2.1
uvicorn==0.20.0
websockets==10.4
Werkzeug==3.0.3
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post0
